+++ CotEditor/Sources/Help.swift
@@ -1,0 +1,56 @@
+/*
+ 
+ Help.swift
+ 
+ CotEditor
+ https://coteditor.com
+ 
+ Created by 1024jp on 2016-10-19.
+ 
+ ------------------------------------------------------------------------------
+ 
+ © 2016 1024jp
+ 
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+ 
+ https://www.apache.org/licenses/LICENSE-2.0
+ 
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ 
+ */
+
+struct Help {
+    
+    static let anchors = [
+        "releasenotes",
+        "pref_general",
+        "pref_window",
+        "pref_appearance",
+        "pref_edit",
+        "pref_format",  // 5
+        "pref_filedrop",
+        "pref_keybindings",
+        "pref_print",
+        "whats_new",
+        "specification_changes",  // 10
+        "howto_customize_scriptmenu",
+        "about_applescript",
+        "about_unixscript",
+        "pref_integration",
+        "about_file_mapping",  // 15
+        "about_cot",
+        "about_syntaxstyle",
+        "about_comment_settings",
+        "about_outlinemenu_settings",
+        "about_complist_settings",  // 20
+        "about_file_mapping",
+        "about_styleinfo_settings",
+        "howto_findpanel",
+        ]
+}